requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
DAWG-Python==0.7.2
docopt==0.6.2
exceptiongroup==1.2.2
fastapi==0.115.5
frozenlist==1.5.0
h11==0.14.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
yarl==1.17.1
httpx
